🚨 Address knip lints
Seems like <script module> is not properly processed by Knip
